Commit 5 - update heath UI and Homing Enemy Implementation

+++ Assets/Scripts/Weapon/Weapon.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Weapon : MonoBehaviour
 {
     [SerializeField]
-    protected int damage;
+    protected float damage;
     [SerializeField]
     protected float coolDown;
